Send SET_ROW to editor as only one byte
All commands in the rocket protocol are only one byte long. Use
writeByte instead of writeInt for Commands.SET_ROW in
controllerRowChanged(). writeInt sends four bytes of which three are
zeros in this case.

Command processing loop in some known editors appear to a) process
commands byte by byte and b) ignore unknown bytes, so this might not
have affected them, but let's be consistent anyway.