Add buffering to input/output TCP streams
This is not really a bugfix; it just greatly reduces the amount of
syscalls required to send commands to Rocket. Flushing is now
absolutely required at command boundaries as otherwise the commands
might never be sent.

In the future the better way to do the whole TCP I/O would be to read
bigger chunks into intermediate buffers and parse commands from there
without blocking if data is not fully available. This would be
beneficial especially at startup when Rocket sends all the data over
the socket.

Big thanks to @bjakke for these suggestions!